pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clan-manager-bot"
version = "0.1.0"
description = "AI-powered clan chat manager bot for Telegram"
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot[job-queue]==22.6",
  "asyncpg==0.30.0",
  "pydantic==2.12.5",
  "pydantic-settings==2.11.0",
  "python-dotenv==1.2.1",
  "orjson==3.11.4",
  "httpx==0.28.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
